Dirty realisation Sessions With Remember_me Option

+++ app/controllers/application_controller.rb
@@ -2,13 +2,19 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  helper_method :sign_in, :current_user, :user_signed_in?
+  helper_method :sign_in, :current_user, :user_signed_in?, :sign_out
   def sign_in(user)
     session[:user_id] = user.id
   end
 
   def current_user
-    @current_user ||= User.find_by(id: session[:user_id])
+    if ( cookies.signed[:user_id])
+       user = User.find_by(id: cookies.signed[:user_id])
+      if  user && user.authenticated?(cookies[:remember_token])
+        sign_in(user)
+      end
+    end
+      @current_user ||= User.find_by(id: session[:user_id])
   end
 
   def user_signed_in?
@@ -21,7 +27,18 @@ class ApplicationController < ActionController::Base
   end
 
   def sign_out
+    return false unless user_signed_in?
+    @current_user.update_attribute(:remember_digest,nil)
+    cookies.delete(:user_id)
+    cookies.delete(:remember_token)
     session.delete(:user_id)
     @current_user = nil
+  end
+
+  def remember_me(user)
+    remember_token = SecureRandom.urlsafe_base64
+    user.update_attribute(:remember_digest, BCrypt::Password.create(remember_token))
+    cookies.permanent.signed[:user_id] = user.id
+    cookies.permanent[:remember_token] = remember_token
   end
 end
